chore: move DotNotationMotionStringToken to types file
